SAM/AI-GoogleGemini/src/requirements.txt
google-generativeai
Pillow
botocore